+++ core/src/cl/makinolas/atk/actors/friend/Igglypuff.java
@@ -23,13 +23,13 @@ public class Igglypuff extends AbstractFriend {
     setMaxMagic(100);
   }
   
-  public Igglypuff(float level){
+  public Igglypuff(int level){
     this();
     initLevel(level);
   }
   
   @Override
-  protected void initLevel(float level){
+  protected void initLevel(int level){
    this.level = new Level(level);
    new Evolution(this.level, 16, 1);
   }
