+++ core/src/cl/makinolas/atk/actors/friend/Mareep.java
@@ -23,13 +23,13 @@ public class Mareep extends AbstractFriend {
     setMaxMagic(100);
   }
   
-  public Mareep(float level){
+  public Mareep(int level){
     this();
     initLevel(level);
   }
   
   @Override
-  protected void initLevel(float level){
+  protected void initLevel(int level){
    this.level = new Level(level);
    new Evolution(this.level, 16, 1);
    new Evolution(this.level, 30, 2);
