+++ core/src/cl/makinolas/atk/actors/friend/Shinx.java
@@ -23,13 +23,13 @@ public class Shinx extends AbstractFriend{
     setMaxMagic(100);
   }
   
-  public Shinx(float level){
+  public Shinx(int level){
     this();
     initLevel(level);
   }
   
   @Override
-  protected void initLevel(float level){
+  protected void initLevel(int level){
    this.level = new Level(level);
    new Evolution(this.level, 16, 1);
    new Evolution(this.level, 32, 1);
