+++ core/src/cl/makinolas/atk/actors/friend/Zubat.java
@@ -28,13 +28,13 @@ public class Zubat extends AbstractFriend {
     setMaxMagic(100);
   }
   
-  public Zubat(float level){
+  public Zubat(int level){
     this();
     initLevel(level);
   }
   
   @Override
-  protected void initLevel(float level){
+  protected void initLevel(int level){
    this.level = new Level(level);
    new Evolution(this.level, 20, 1);
    new Evolution(this.level, 30, 2);
